backend/requirements.txt
djangorestframework
psycopg2-binary
python-dotenv
djangorestframework-simplejwt